Exception handling - delete

+++ src/main/java/com/webServicesProjects/Web/_Services_Spring_Boot_JPA_Hibernate/services/UserService.java
@@ -2,8 +2,11 @@ package com.webServicesProjects.Web._Services_Spring_Boot_JPA_Hibernate.services
 
 import com.webServicesProjects.Web._Services_Spring_Boot_JPA_Hibernate.entities.User;
 import com.webServicesProjects.Web._Services_Spring_Boot_JPA_Hibernate.repositories.UserRepository;
+import com.webServicesProjects.Web._Services_Spring_Boot_JPA_Hibernate.services.exceptions.DatabaseException;
 import com.webServicesProjects.Web._Services_Spring_Boot_JPA_Hibernate.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -28,7 +31,17 @@ public class UserService {
     }
 
     public void delete(Long id){
-        userRepository.deleteById(id);
+        try {
+            if (userRepository.existsById(id)) {
+                userRepository.deleteById(id);
+            } else {
+                throw new ResourceNotFoundException(id);
+            }
+        } catch (ResourceNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj){
